+++ android/src/main/kotlin/com/kasem/receive_sharing_intent/ReceiveSharingIntentPlugin.kt
@@ -27,7 +27,6 @@ import java.net.URLConnection
 private const val MESSAGES_CHANNEL = "receive_sharing_intent/messages"
 private const val EVENTS_CHANNEL_MEDIA = "receive_sharing_intent/events-media"
 private const val EVENTS_CHANNEL_TEXT = "receive_sharing_intent/events-text"
-private const val EVENTS_CHANNEL_LINK = "receive_sharing_intent/events-link"
 
 class ReceiveSharingIntentPlugin : FlutterPlugin, ActivityAware, MethodCallHandler,
         EventChannel.StreamHandler, NewIntentListener {
@@ -38,12 +37,8 @@ class ReceiveSharingIntentPlugin : FlutterPlugin, ActivityAware, MethodCallHandl
     private var initialText: String? = null
     private var latestText: String? = null
 
-    private var initialLink: String? = null
-    private var latestLink: String? = null
-
     private var eventSinkMedia: EventChannel.EventSink? = null
     private var eventSinkText: EventChannel.EventSink? = null
-    private var eventSinkLink: EventChannel.EventSink? = null
 
     private var binding: ActivityPluginBinding? = null
     private lateinit var applicationContext: Context
@@ -57,9 +52,6 @@ class ReceiveSharingIntentPlugin : FlutterPlugin, ActivityAware, MethodCallHandl
 
         val eChannelText = EventChannel(binaryMessenger, EVENTS_CHANNEL_TEXT)
         eChannelText.setStreamHandler(this)
-
-        val eChannelLink = EventChannel(binaryMessenger, EVENTS_CHANNEL_LINK)
-        eChannelLink.setStreamHandler(this)
     }
 
     override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
@@ -74,7 +66,6 @@ class ReceiveSharingIntentPlugin : FlutterPlugin, ActivityAware, MethodCallHandl
         when (arguments) {
             "media" -> eventSinkMedia = events
             "text" -> eventSinkText = events
-            "link" -> eventSinkLink = events
         }
     }
 
@@ -82,7 +73,6 @@ class ReceiveSharingIntentPlugin : FlutterPlugin, ActivityAware, MethodCallHandl
         when (arguments) {
             "media" -> eventSinkMedia = null
             "text" -> eventSinkText = null
-            "link" -> eventSinkLink = null
         }
     }
 
@@ -109,14 +99,11 @@ class ReceiveSharingIntentPlugin : FlutterPlugin, ActivityAware, MethodCallHandl
         when (call.method) {
             "getInitialMedia" -> result.success(initialMedia?.toString())
             "getInitialText" -> result.success(initialText)
-            "getInitialLink" -> result.success(initialLink)
             "reset" -> {
                 initialMedia = null
                 latestMedia = null
                 initialText = null
                 latestText = null
-                initialLink = null
-                latestLink = null
                 result.success(null)
             }
             else -> result.notImplemented()
@@ -143,9 +130,9 @@ class ReceiveSharingIntentPlugin : FlutterPlugin, ActivityAware, MethodCallHandl
             }
             intent.action == Intent.ACTION_VIEW -> { // Opening URL
                 val value = intent.dataString
-                if (initial) initialLink = value
-                latestLink = value
-                eventSinkLink?.success(latestLink)
+                if (initial) initialText = value
+                latestText = value
+                eventSinkText?.success(latestText)
             }
         }
     }
@@ -156,7 +143,7 @@ class ReceiveSharingIntentPlugin : FlutterPlugin, ActivityAware, MethodCallHandl
         return when (intent.action) {
             Intent.ACTION_SEND -> {
                 val uri = intent.getParcelableExtra<Uri>(Intent.EXTRA_STREAM)
-                val path = if (uri != null) FileDirectory.getAbsolutePath(applicationContext, uri) else null
+                val path = uri?.let{ FileDirectory.getAbsolutePath(applicationContext, it) }
                 if (path != null) {
                     val type = getMediaType(path)
                     val thumbnail = getThumbnail(path, type)